refactor(cli): extract history.py command module from cli monolith
Fourth slice of the cli monolith decomposition. Moves the history and
report commands, their display helpers (_parse_lookback_days,
_display_scan_event_row, _display_scan_history), and all audit-event /
audit-chain / history constants from phi_scan/cli/__init__.py into
phi_scan/cli/history.py.

Commands are re-registered via app.command(name)(function). Four test
monkeypatch targets for phi_scan.cli.DEFAULT_DATABASE_PATH used by the
report and history commands now point at
phi_scan.cli.history.DEFAULT_DATABASE_PATH (the dashboard test keeps
the phi_scan.cli target since dashboard is unmoved). No behavior
change.

cli/__init__.py shrinks 1068 → 952 lines.
