+++ phi_scan/cli/__init__.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import hashlib
 import json
 import logging
 import time
@@ -16,12 +15,10 @@ from watchdog.observers import Observer
 
 from phi_scan import __version__
 from phi_scan.audit import (
-    ChainVerifyResult,
     ensure_current_schema,
     get_last_scan,
     insert_scan_event,
     query_recent_scans,
-    verify_audit_chain,
 )
 from phi_scan.ci_integration import (
     CIIntegrationError,
@@ -94,7 +91,6 @@ from phi_scan.constants import (
     DEFAULT_DATABASE_PATH,
     EXIT_CODE_CLEAN,
     EXIT_CODE_ERROR,
-    EXIT_CODE_VIOLATION,
     OutputFormat,
 )
 from phi_scan.exceptions import (
@@ -228,50 +224,6 @@ _FRAMEWORK_PARSE_ERROR: str = "Invalid --framework value: {error}"
 _WATCH_PATH_HELP: str = "Directory to watch for file system changes."
 
 # ---------------------------------------------------------------------------
-# History command
-# ---------------------------------------------------------------------------
-
-_HISTORY_LAST_HELP: str = "Show scans from the last N days (e.g. 30d)."
-_HISTORY_VERIFY_HELP: str = (
-    "Recompute HMAC-SHA256 hash chain and report PASS or FAIL. "
-    "Exits with code 1 if the chain is broken (tamper detected)."
-)
-_HISTORY_REPO_HELP: str = (
-    "Filter by repository path (e.g. /home/user/my-repo). "
-    "The path is SHA-256 hashed before comparison against stored repository_hash values."
-)
-_HISTORY_VIOLATIONS_ONLY_HELP: str = (
-    "Show only scans where PHI findings were detected (is_clean=false)."
-)
-_DEFAULT_HISTORY_PERIOD: str = "30d"
-_DAYS_PERIOD_SUFFIX: str = "d"
-_HISTORY_PERIOD_FORMAT_ERROR: str = (
-    "Period must be in the format '30d' (number of days), got {period!r}"
-)
-_NO_SCAN_HISTORY_MESSAGE: str = "No scan history found."
-_HISTORY_ROW_FORMAT: str = "{scanned_at}  {status}  risk={risk_level}  files={files_scanned}"
-_ZERO_FILES_SCANNED: int = 0
-
-# ---------------------------------------------------------------------------
-# Report command
-# ---------------------------------------------------------------------------
-
-_NO_LAST_SCAN_MESSAGE: str = "No scan record found. Run `phi-scan scan` first."
-_LAST_SCAN_HEADER: str = "Last scan result:"
-
-# ---------------------------------------------------------------------------
-# Audit event dict keys (matching column names in the audit SQLite schema)
-# ---------------------------------------------------------------------------
-
-_AUDIT_KEY_SCANNED_AT: str = "scanned_at"
-_AUDIT_KEY_IS_CLEAN: str = "is_clean"
-_AUDIT_KEY_RISK_LEVEL: str = "risk_level"
-_AUDIT_KEY_FILES_SCANNED: str = "files_scanned"
-_CLEAN_STATUS_LABEL: str = "CLEAN"
-_VIOLATION_STATUS_LABEL: str = "VIOLATION"
-_UNKNOWN_LABEL: str = "unknown"
-
-# ---------------------------------------------------------------------------
 # Dashboard
 # ---------------------------------------------------------------------------
 
@@ -325,20 +277,6 @@ _AUDIT_KEY_MISSING_DEBUG: str = (
 )
 _NOTIFICATION_EMAIL_FAILURE_WARNING: str = "Email notification failed: {error}"
 _NOTIFICATION_WEBHOOK_FAILURE_WARNING: str = "Webhook notification failed: {error}"
-_AUDIT_CHAIN_PASS_MESSAGE: str = "Audit chain integrity: PASS — all row hashes verified."
-_AUDIT_CHAIN_FAIL_MESSAGE: str = (
-    "Audit chain integrity: FAIL — one or more rows failed hash verification. "
-    "The audit log may have been tampered with."
-)
-_AUDIT_CHAIN_SKIP_MESSAGE: str = (
-    "Audit chain verification skipped — no audit key found. "
-    "Run 'phi-scan setup' to generate the key."
-)
-_AUDIT_CHAIN_SKIPPED_ROWS_WARNING: str = (
-    "Warning: {skipped_rows} row(s) had no chain hash and were not verified. "
-    "Treat this audit as partially unverified."
-)
-_AUDIT_CHAIN_VERIFY_FLAG: str = "--verify"
 _SPINNER_NOTIFY_MESSAGE: str = "Sending notifications…"
 
 # ---------------------------------------------------------------------------
@@ -583,65 +521,6 @@ def _prepare_scan_phase(
         display_phase_scanning()
     emit_verbose_phase(_VERBOSE_PHASE_SCANNING.format(count=len(scan_targets)), is_verbose)
     return scan_targets
-
-
-# ---------------------------------------------------------------------------
-# Internal helpers — history and report commands
-# ---------------------------------------------------------------------------
-
-
-def _parse_lookback_days(period: str) -> int:
-    """Parse a period string like '30d' into an integer number of days.
-
-    Args:
-        period: A string ending in 'd' with a positive integer prefix.
-
-    Returns:
-        Number of lookback days as an integer.
-
-    Raises:
-        typer.BadParameter: If period is not in the expected format.
-    """
-    if not period.endswith(_DAYS_PERIOD_SUFFIX):
-        raise typer.BadParameter(_HISTORY_PERIOD_FORMAT_ERROR.format(period=period))
-    day_count_str = period[: -len(_DAYS_PERIOD_SUFFIX)]
-    if not day_count_str.isdigit():
-        raise typer.BadParameter(_HISTORY_PERIOD_FORMAT_ERROR.format(period=period))
-    return int(day_count_str)
-
-
-def _display_scan_event_row(scan_event_record: dict[str, Any]) -> None:
-    """Print a single audit scan event as a one-line summary.
-
-    Args:
-        scan_event_record: Audit row dict as returned by get_last_scan or query_recent_scans.
-    """
-    scanned_at = scan_event_record.get(_AUDIT_KEY_SCANNED_AT, _UNKNOWN_LABEL)
-    is_clean = scan_event_record.get(_AUDIT_KEY_IS_CLEAN, False)
-    risk_level = scan_event_record.get(_AUDIT_KEY_RISK_LEVEL, _UNKNOWN_LABEL)
-    files_scanned = scan_event_record.get(_AUDIT_KEY_FILES_SCANNED, _ZERO_FILES_SCANNED)
-    status = _CLEAN_STATUS_LABEL if is_clean else _VIOLATION_STATUS_LABEL
-    typer.echo(
-        _HISTORY_ROW_FORMAT.format(
-            scanned_at=scanned_at,
-            status=status,
-            risk_level=risk_level,
-            files_scanned=files_scanned,
-        )
-    )
-
-
-def _display_scan_history(scan_events: list[dict[str, Any]]) -> None:
-    """Print a list of audit scan events, or a no-history message if empty.
-
-    Args:
-        scan_events: List of audit row dicts from query_recent_scans.
-    """
-    if not scan_events:
-        typer.echo(_NO_SCAN_HISTORY_MESSAGE)
-        return
-    for scan_event in scan_events:
-        _display_scan_event_row(scan_event)
 
 
 # ---------------------------------------------------------------------------
@@ -936,62 +815,7 @@ def watch(
         observer.join()
 
 
-@app.command("report")
-def display_last_scan() -> None:
-    """Display the most recent scan result from the audit log."""
-    database_path = Path(DEFAULT_DATABASE_PATH).expanduser()
-    ensure_current_schema(database_path)
-    last_scan_event = get_last_scan(database_path)
-    if last_scan_event is None:
-        typer.echo(_NO_LAST_SCAN_MESSAGE)
-        return
-    typer.echo(_LAST_SCAN_HEADER)
-    _display_scan_event_row(last_scan_event)
-
-
-@app.command("history")
-def display_history(
-    last: Annotated[str, typer.Option("--last", help=_HISTORY_LAST_HELP)] = _DEFAULT_HISTORY_PERIOD,
-    should_verify: Annotated[
-        bool, typer.Option(_AUDIT_CHAIN_VERIFY_FLAG, help=_HISTORY_VERIFY_HELP)
-    ] = False,
-    repository_path: Annotated[str | None, typer.Option("--repo", help=_HISTORY_REPO_HELP)] = None,
-    should_show_violations_only: Annotated[
-        bool, typer.Option("--violations-only", help=_HISTORY_VIOLATIONS_ONLY_HELP)
-    ] = False,
-) -> None:
-    """Query the audit log for recent scan history."""
-    lookback_days = _parse_lookback_days(last)
-    database_path = Path(DEFAULT_DATABASE_PATH).expanduser()
-    ensure_current_schema(database_path)
-    if should_verify:
-        verify_result: ChainVerifyResult = verify_audit_chain(database_path)
-        if not verify_result.key_present:
-            typer.echo(_AUDIT_CHAIN_SKIP_MESSAGE, err=True)
-        elif verify_result.is_intact:
-            typer.echo(_AUDIT_CHAIN_PASS_MESSAGE)
-            if verify_result.skipped_rows > 0:
-                typer.echo(
-                    _AUDIT_CHAIN_SKIPPED_ROWS_WARNING.format(
-                        skipped_rows=verify_result.skipped_rows
-                    ),
-                    err=True,
-                )
-        else:
-            typer.echo(_AUDIT_CHAIN_FAIL_MESSAGE, err=True)
-            raise typer.Exit(code=EXIT_CODE_VIOLATION)
-    repository_hash = (
-        hashlib.sha256(repository_path.encode("utf-8")).hexdigest() if repository_path else None
-    )
-    scan_events = query_recent_scans(
-        database_path,
-        lookback_days,
-        repository_hash=repository_hash,
-        should_show_violations_only=should_show_violations_only,
-    )
-    _display_scan_history(scan_events)
-
-
+from phi_scan.cli.history import display_history, display_last_scan  # noqa: E402
 from phi_scan.cli.hooks import (  # noqa: E402
     download_models,
     initialize_project,
@@ -999,6 +823,8 @@ from phi_scan.cli.hooks import (  # noqa: E402
     uninstall_hook,
 )
 
+app.command("report")(display_last_scan)
+app.command("history")(display_history)
 app.command("install-hook")(install_hook)
 app.command("uninstall-hook")(uninstall_hook)
 app.command("init")(initialize_project)
